feat(healthevent, healthcomponent, healthui): modifying health player logic and asset
healthEvent: was added more events inside healthEvents for better logic: healthChanged,
maxHealthChanged and healthInitialized. healthComponent: implemetation of a timeOut for
globalEventEmitter, method of increaseHealth responsible for increase player health after buy a
item, and loseHealth was incremeted with a if condition to prevent error. HealthUi: was added a new
sprite/image of the healthbar and all the logic responsible for losing health and automatically crop
the healthbar, as the handleHealthUpdate responsible for increase the maxHealth.